requirements.txt
# Telegram Bot
python-telegram-bot==22.5

# AI & Face Recognition
insightface==0.7.3
onnxruntime==1.23.2
faiss-cpu==1.12.0

# Image Processing
opencv-python==4.12.0.88
Pillow==10.1.0
numpy==2.2.6

# Database
SQLAlchemy==2.0.44
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0
